+++ libfreerdp/core/nla.c
@@ -1244,7 +1244,7 @@ int credssp_recv(rdpCredssp* credssp)
 			!ber_read_sequence_tag(s, &length) || /* NegoDataItem */
 			!ber_read_contextual_tag(s, 0, &length, TRUE) || /* [0] negoToken */
 			!ber_read_octet_string_tag(s, &length) || /* OCTET STRING */
-			Stream_GetRemainingLength(s) < length)
+			((int) Stream_GetRemainingLength(s)) < length)
 		{
 			Stream_Free(s, TRUE);
 			return -1;
@@ -1257,8 +1257,8 @@ int credssp_recv(rdpCredssp* credssp)
 	/* [2] authInfo (OCTET STRING) */
 	if (ber_read_contextual_tag(s, 2, &length, TRUE) != FALSE)
 	{
-		if(!ber_read_octet_string_tag(s, &length) || /* OCTET STRING */
-			Stream_GetRemainingLength(s) < length)
+		if (!ber_read_octet_string_tag(s, &length) || /* OCTET STRING */
+			((int) Stream_GetRemainingLength(s)) < length)
 		{
 			Stream_Free(s, TRUE);
 			return -1;
@@ -1271,8 +1271,8 @@ int credssp_recv(rdpCredssp* credssp)
 	/* [3] pubKeyAuth (OCTET STRING) */
 	if (ber_read_contextual_tag(s, 3, &length, TRUE) != FALSE)
 	{
-		if(!ber_read_octet_string_tag(s, &length) || /* OCTET STRING */
-			Stream_GetRemainingLength(s) < length)
+		if (!ber_read_octet_string_tag(s, &length) || /* OCTET STRING */
+			((int) Stream_GetRemainingLength(s)) < length)
 		{
 			Stream_Free(s, TRUE);
 			return -1;
